+++ iphone/Classes/TiUIiOSCoverFlowView.h
@@ -18,7 +18,7 @@
 	NSMutableDictionary* toLoad;
 	NSMutableDictionary* loading;
 	NSRecursiveLock* loadLock;
-	int previous;
+	NSInteger previous;
 }
 
 -(void)setImage:(id)image forIndex:(NSInteger)index;
